+++ morphir/runtime/src/org/finos/morphir/runtime/Distributions.scala
@@ -20,22 +20,6 @@ import org.finos.morphir.ir.distribution.Distribution
 import org.finos.morphir.ir.distribution.Distribution.Library
 import zio.Chunk
 
-//TODO: Error hierarchy and code should reflect possibility of specification lookup
-sealed abstract class LookupError(msg: String) extends Exception(msg) {
-  def getMsg: String = msg
-}
-case class MissingPackage(pkgName: PackageName) extends LookupError(s"Package ${pkgName.toString} not found")
-case class MissingModule(pkgName: PackageName, modName: ModuleName)
-    extends LookupError(s"Package ${pkgName.toString} does not contain module ${modName.toString}")
-case class MissingType(pkgName: PackageName, modName: ModuleName, typeName: Name)
-    extends LookupError(s"Module ${pkgName.toString}:${modName.toString} has no type named ${typeName.toTitleCase}")
-case class MissingDefinition(pkgName: PackageName, modName: ModuleName, defName: Name) extends LookupError(
-      s"Module ${pkgName.toString}:${modName.toString} has no definition named ${defName.toCamelCase}"
-    )
-case class MissingConstructor(pkgName: PackageName, modName: ModuleName, ctorName: Name) extends LookupError(
-      s"Module ${pkgName.toString}:${modName.toString} has no constructor named ${ctorName.toTitleCase}"
-    )
-
 class Distributions(dists: Map[PackageName, Distribution]) {
   def lookupModuleSpecification(pkgName: PackageName, modName: ModuleName): Either[LookupError, ModSpec.Raw] =
     dists.get(pkgName) match {
