+++ morphir/runtime/src/org/finos/morphir/runtime/TypedMorphirRuntime.scala
@@ -14,8 +14,8 @@ import org.finos.morphir.runtime.environment.MorphirEnv
 import org.finos.morphir.runtime.exports.*
 
 trait TypedMorphirRuntime extends MorphirRuntime {
-  type TypeAttribs  = TypedMorphirRuntime.TypeAttribs
-  type ValueAttribs = TypedMorphirRuntime.ValueAttribs
+  type TypeAttribs  = TypedMorphirRuntimeDefs.TypeAttribs
+  type ValueAttribs = TypedMorphirRuntimeDefs.ValueAttribs
 
   final def evaluate(
       entryPoint: Value[scala.Unit, UType],
@@ -44,11 +44,4 @@ trait TypedMorphirRuntime extends MorphirRuntime {
     val inputIRs = params.map(toValue(_))
     evaluate(entryPoint, inputIR, inputIRs: _*)
   }
-}
-
-object TypedMorphirRuntime {
-  final type TypeAttribs  = scala.Unit
-  final type ValueAttribs = UType
-  type RuntimeValue       = Value[TypeAttribs, ValueAttribs]
-  type RuntimeDefinition  = Value.Definition[TypeAttribs, ValueAttribs]
 }
