--- morphir/runtime/src/org/finos/morphir/runtime/package.scala
@@ -1,4 +1,0 @@
-package org.finos.morphir
-import zio.prelude.fx.ZPure
-
-package object runtime {}
