--- morphir/src/morphir/sdk/LocalTime.scala
@@ -1,3 +1,0 @@
-package morphir.sdk
-
-object LocalTime {}
