--- morphir/src/morphir/sdk/Maybe.scala
@@ -1,3 +1,0 @@
-package morphir.sdk
-
-object Maybe {}
