--- morphir/src/morphir/sdk/String.scala
@@ -1,8 +1,0 @@
-package morphir.sdk
-
-/// Utilities to extend the basic String operations provided by `elm/core`.
-object String {
-  type String = scala.Predef.String
-
-  // def ofLength[A](length: Basics.Int) = ???
-}
