+++ morphir/src/org/finos/morphir/ModuleDescriptor.scala
@@ -1,4 +1,4 @@
-package org.finos.morphir.runtime.quick
+package org.finos.morphir
 
 import org.finos.morphir.naming._
 import org.finos.morphir.MorphirTag
