+++ morphir/src/org/finos/morphir/runtime/environment.scala
@@ -1,7 +1,7 @@
 package org.finos.morphir.runtime
 import org.finos.morphir.runtime.services.kernel.MorphirKernel
 import org.finos.morphir.runtime.services.sdk.MorphirSdk
-import zio._
+import zio.*
 
 object environment {
   type MorphirEnv = MorphirSdk with MorphirKernel
