--- morphir/src/org/finos/morphir/runtime/services/sdk/LocalTimeModule.scala
@@ -1,3 +1,0 @@
-package org.finos.morphir.runtime.services.sdk
-
-object LocalTimeModule {}
